requirements.txt
pandas>=1.5.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0
python-dotenv>=1.0.0
selenium>=4.15.0
openpyxl>=3.1.0
html2text>=2020.1.16
jinja2>=3.1.0